backend/test-requirements.txt
# Core testing frameworks
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# HTTP client for testing APIs
httpx==0.25.2

# Application dependencies for testing
fastapi[all]==0.104.1
pydantic==2.5.0
bcrypt==4.0.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
